Add deleted date in data model
