+++ src/main/java/fr/openent/competences/Utils.java
@@ -18,7 +18,6 @@
 package fr.openent.competences;
 
 import fr.openent.competences.bean.Eleve;
-import fr.openent.competences.constants.Field;
 import fr.openent.competences.service.MatiereService;
 import fr.openent.competences.service.UtilsService;
 import fr.openent.competences.service.impl.DefaultMatiereService;
@@ -102,50 +101,6 @@ public class Utils {
         }));
     }
 
-    public static JsonArray filterSubtitute(List<Object> periodes, JsonArray multiTeachers) {
-        for (Object periodeO : periodes) {
-            JsonObject periode = (JsonObject) periodeO;
-            multiTeachers = new JsonArray(multiTeachers.stream().filter(obj -> {
-                JsonObject multi = (JsonObject) obj;
-                if (!multi.getBoolean(Field.IS_COTEACHING)) {
-                    String classOrGroupId =
-                            (periode.containsKey(Field.ID_CLASSE))
-                                    ? periode.getString(Field.ID_CLASSE) : periode.getString(Field.ID_GROUPE);
-                    return filterSubtituteByDatePeriode(periode, multi, classOrGroupId);
-                } else {
-                    return true;
-                }
-            }).collect(Collectors.toList()));
-        }
-        return multiTeachers;
-    }
-
-    private static boolean filterSubtituteByDatePeriode(JsonObject periode, JsonObject multi, String classOrGroupId) {
-        if (classOrGroupId.equals(multi.getString(Field.CLASS_OR_GROUP_ID))) {
-            SimpleDateFormat formatter1 = new SimpleDateFormat(Field.dateFormateYYYYMMDDTHHMMSS);
-            Date multiStartDate;
-            Date multiEndDate;
-            Date periodeStartDate;
-            Date periodeEndDate;
-            try {
-                multiStartDate = formatter1.parse(multi.getString(Field.START_DATE));
-                multiEndDate = formatter1.parse(multi.getString(Field.END_DATE));
-                periodeStartDate = formatter1.parse(periode.getString(Field.TIMESTAMP_DT));
-                periodeEndDate = formatter1.parse(periode.getString(Field.TIMESTAMP_FN));
-            } catch (ParseException e) {
-                log.error("[Competences@FilterSubtitute] cannot parse dates");
-                return true;
-            }
-            if (multiStartDate != null && multiEndDate != null && periodeEndDate != null && periodeStartDate != null)
-                return (multiStartDate.after(periodeStartDate) && multiStartDate.before(periodeEndDate))
-                        || (multiEndDate.after(periodeStartDate) && multiEndDate.before(periodeEndDate));
-            else {
-                return true;
-            }
-        } else {
-            return true;
-        }
-    }
 
     /**
      * retourne une classe avec ses groups (ids)
